+++ desert-core/src/test/scala/io/github/vigoo/desert/DefaultTypeRegistrySpec.scala
@@ -1,11 +1,10 @@
 package io.github.vigoo.desert
 
-import io.github.vigoo.desert.TypeRegistry.RegisteredTypeId
 import io.github.vigoo.desert.codecs._
+import io.github.vigoo.desert.TypeRegistry.RegisteredTypeId
 import zio.test.Assertion._
 import zio.test._
 import zio.test.environment.TestEnvironment
-
 
 object DefaultTypeRegistrySpec extends DefaultRunnableSpec {